requirements.txt
django
djangorestframework
django-cors-headers
django-json-widget
djangorestframework-simplejwt
jsonschema
pydantic

psycopg2-binary==2.9.3
